Move to public API
